FEAT: Dominant Key feature parse
Added functionality to extract the dominant key from a musicxml file using
David Temperley's Key Profiles.
Uses The normalized score notes' frequencies and compares to the
key profiles using euclidean distance, the closest one is determined to
be the key profile. The main key note is obtained by the maximum weighted
note in the frequencies list.